Move `newMessage` package outside of `main` package
