+++ app/src/main/java/com/infomaniak/mail/ui/newMessage/RecipientFieldView.kt
@@ -15,7 +15,7 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package com.infomaniak.mail.ui.main.newMessage
+package com.infomaniak.mail.ui.newMessage
 
 import android.content.Context
 import android.content.res.ColorStateList
