+++ app/src/main/res/layout/activity_new_message.xml
@@ -20,7 +20,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.main.newMessage.NewMessageActivity">
+    tools:context=".ui.newMessage.NewMessageActivity">
 
     <LinearLayout
         android:layout_width="match_parent"
