+++ app/src/main/res/layout/fragment_ai_prompt.xml
@@ -21,7 +21,7 @@
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
     android:theme="@style/AiCursorAndPrimaryColorTheme"
-    tools:context=".ui.main.newMessage.AiPromptFragment">
+    tools:context=".ui.newMessage.AiPromptFragment">
 
     <com.google.android.material.card.MaterialCardView
         style="@style/AiCardView"
