+++ app/src/main/res/layout/view_recipient_field.xml
@@ -74,7 +74,7 @@
         app:layout_constraintStart_toEndOf="@id/prefix"
         app:layout_constraintTop_toBottomOf="@id/chipsRecyclerView">
 
-        <com.infomaniak.mail.ui.main.newMessage.BackspaceAwareTextInput
+        <com.infomaniak.mail.ui.newMessage.BackspaceAwareTextInput
             android:id="@+id/textInput"
             android:layout_width="match_parent"
             android:layout_height="match_parent"
